# contracts/openapi.engine.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scrapix API",
    "description": "High-performance web crawler and search indexer API. Scrape pages, map websites, run distributed crawls, and search indexed content.",
    "contact": {
      "name": "Meilisearch",
      "url": "https://scrapix.meilisearch.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://scrapix.meilisearch.dev",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "paths": {
    "/analytics/v0/pipes": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "List the available pipes and their parameters.",
        "operationId": "list_pipes",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PipeInfo"
                  }
                }
              }
            }
          },
          "404": {
            "description": "ClickHouse is not configured, or `account_id` names another account"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/analytics/v0/pipes/kpis.json": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Key performance indicators, aggregated from `top_domains` (limit 10 000).",
        "operationId": "kpis",
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "description": "Look-back window in hours (default 24)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "account_id",
            "in": "query",
            "description": "Account filter; hosted callers may only name their own account",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tinybird envelope: meta, data, rows, statistics"
          },
          "404": {
            "description": "ClickHouse is not configured, or `account_id` names another account"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/batch/scrape": {
      "post": {
        "tags": [
          "scrape"
        ],
        "summary": "Scrape many URLs as one job",
        "description": "Starts a job that runs the `/scrape` pipeline for every URL (bounded\nconcurrency) and returns its `job_id` immediately. Track it with\n`GET /job/{id}/status`, `/job/{id}/events` or webhooks, read the pages\nwith `GET /job/{id}/results`, cancel with `DELETE /job/{id}`. A URL that\nfails is reported as a result with `success: false` and an `error`; it\ndoes not fail the batch. Credits: same as `/scrape`, per URL scraped.",
        "operationId": "batch_scrape",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchScrapeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchScrapeResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "402": {
            "description": "Not enough credits for the whole batch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/crawl": {
      "post": {
        "tags": [
          "crawl"
        ],
        "summary": "Create a new async crawl job",
        "operationId": "create_crawl",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrawlConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCrawlResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/crawl/bulk": {
      "post": {
        "tags": [
          "crawl"
        ],
        "summary": "Create multiple crawl jobs from a batch of configs",
        "operationId": "create_crawl_bulk",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CrawlConfig"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkCrawlResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/crawl/sync": {
      "post": {
        "tags": [
          "crawl"
        ],
        "summary": "Create a sync crawl job (waits for completion)",
        "description": "With `include_results=true`, the response also carries the first page of\nthe job's results (`GET /job/{id}/results`).",
        "operationId": "create_crawl_sync",
        "parameters": [
          {
            "name": "include_results",
            "in": "query",
            "description": "Also return the first page of the job's results (default false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "results_limit",
            "in": "query",
            "description": "Page size of the included results (default 20, max 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrawlConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlSyncResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/domains": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Domains endpoint",
        "operationId": "handle_domains",
        "parameters": [
          {
            "name": "top",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/errors": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Errors endpoint",
        "operationId": "handle_errors",
        "parameters": [
          {
            "name": "last",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "job_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/extract": {
      "post": {
        "tags": [
          "extract"
        ],
        "summary": "Extract structured data from web pages",
        "description": "Starts a job that fetches every URL (globs such as\n`https://example.com/blog/*` are resolved first, 50 URLs max), then runs\none AI extraction over their content following `prompt` and/or `schema`.\nPoll `GET /extract/{id}` for the result; the fetched pages are available\nfrom `GET /job/{id}/results`. Requires an AI provider on the server.",
        "operationId": "create_extract",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateExtractResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "402": {
            "description": "Not enough credits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "No AI provider is configured on the server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/extract/{id}": {
      "get": {
        "tags": [
          "extract"
        ],
        "summary": "Get an extraction's status and result",
        "operationId": "get_extract",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Extract job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/job/{id}": {
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel a job",
        "description": "Stops the job everywhere (frontier and workers) and bills the pages\ncrawled so far. Only a pending, running or paused job can be cancelled:\na job that already completed, failed or was cancelled returns 409 and\nkeeps its status.",
        "operationId": "cancel_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The job is already terminal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/job/{id}/resume": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Resume a paused job",
        "description": "Only a paused job can be resumed; any other status returns 409.",
        "operationId": "resume_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The job is not paused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchScrapeResponse": {
        "type": "object",
        "description": "Response of `POST /batch/scrape`",
        "required": [
          "job_id",
          "status",
          "urls_count",
          "message"
        ],
        "properties": {
          "job_id": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "urls_count": {
            "type": "integer",
            "description": "Number of URLs accepted",
            "minimum": 0
          }
        }
      },
...
      "PipeInfo": {
        "type": "object",
        "required": [
          "name",
          "description",
          "parameters",
          "endpoint"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "endpoint": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParamInfo"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "security": [
    {
      "api_key": []
    }
  ],
...
}